restructure polyfill.js
