+++ functions/clz32.js
@@ -1,0 +1,6 @@
+module.exports = function clz32 (value) {
+    "use strict";
+
+    value = Number(value) >>> 0;
+    return value ? 32 - value.toString(2).length : 32;
+};
