+++ functions/log10.js
@@ -1,0 +1,5 @@
+module.exports = function log10 (value) {
+    "use strict";
+
+    return Math.log(value) / Math.LN10;
+};
