+++ functions/log1p.js
@@ -1,0 +1,11 @@
+var isNegativeZero = require('./isNegativeZero');
+
+module.exports = function log1p (value) {
+    "use strict";
+
+    if (isNegativeZero(value)) {
+        return value;
+    } else {
+        return Math.log(1 + value);
+    }
+};
