+++ functions/log2.js
@@ -1,0 +1,5 @@
+module.exports = function log2 (value) {
+    "use strict";
+
+    return Math.log(value) / Math.LN2;
+};
