+++ functions/sign.js
@@ -1,0 +1,11 @@
+module.exports = function sign (value) {
+    "use strict";
+
+    if (isNaN(value)) {
+        return NaN;
+    } else if (value === 0) {
+        return value;
+    } else {
+        return value > 0 ? 1 : -1;
+    }
+};
