+++ functions/trunc.js
@@ -1,0 +1,5 @@
+module.exports = function trunc (value) {
+    "use strict";
+
+    return value < 0 ? Math.ceil(value) : Math.floor(value);
+};
